Docker Compose project to regenerate JAXB files
Previously, the files were generated with local versions of Trang/Maven.
It was causing to drown actual changes in cosmetic differences.

This patch replaces all the prerequisites with just Docker Compose,
and allows to regenerate files with the same versions for everyone.

It stores Bash history and Maven cache in local files for convenience.

It refers to MB Solr for making it available through Maven locally,
as we don’t publish this binding in Maven central repository anyway.