MBS-14053: Add ended flag to aliases
This adds "ended" as an attribute to aliases in XML to match what
we already do in JSON.
While ideally we would probably add a life-span element to aliases,
that would be a big change given we already do dates as attributes,
so it's probably better left for ws/3.
For now, this matches what we do for the "primary" boolean.